requirements.txt
pillow==10.3.0
pyfastnoiselite==0.0.5
numpy==2.0.0
imageio==2.34.1